Set time to zero if under a minute and user subtracts a minute
This allows the user to end the half early. Before this patch,
you'd need to hit the minus-second button a bunch of times to get
the game clock down to zero. Now you can just hit the minus-minute
button and both minutes and seconds will zero out. Then hitting
SUBMIT will immediately ring the buzzer and end the half.